Refactor ArrivalMovementService.
Move out the readers for that work across all message types to a
separate object and simplify this service object to only be
concerned with creating MovementMessage and Arrivals.
